decouple custom folder location operation
